requirements.txt
OpenAI-whisper
pandas